% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_housekeep.R
\name{batch_housekeep}
\alias{batch_housekeep}
\title{Found housekeep genes in batch effects group}
\usage{
batch_housekeep(data, batch = NULL, expcv = 1, exphigh = 1, explow = 0)
}
\arguments{
\item{data}{Genomic measure matrix (dimensions probe x sample) - for example, expression matrix}

\item{batch}{must be a vctor,{Batch covariate (only one batch allowed)}}

\item{expcv}{a number range 0 to 1, Sets the maximum quantile value of the coefficient of variation}

\item{exphigh}{a number range 0 to 1, Sets the maximum quantile value of expression sum}

\item{explow}{a number range 0 to 1, Sets the minimum quantile value of expression sum}
}
\value{
data A housekeep gene x sample genomic measure matrix
}
\description{
batch_housekeep allows users to  Found housekeep genes in batch effects group
by coefficient of variation and the expression sum.
}
\examples{
y <- matrix( rpois(1200, lambda=6), nrow=200 )
n = dim(y)
colnames(y) = paste(rep("sample",n[[2]]),1:n[[2]])
rownames(y) = paste(rep("gene",n[[1]]),1:n[[1]])
y = as.data.frame(y)
y[3,] <- c(0,0,0,1,1,1)
y[19,] <- c(1,0,1,0,1,0)
data = y
batch = rep(1:2, each = 3)
#re = batch_housekeep(data, bacth, expcv = 1, exphigh, explow = 0)
}
