% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_cv.R
\name{col_cv}
\alias{col_cv}
\title{Title Calculating coefficient of variation}
\usage{
col_cv(data)
}
\arguments{
\item{data}{must be a data.frame or matrix}
}
\value{
vcxtor
}
\description{
Title Calculating coefficient of variation
}
\examples{
y <- matrix( rpois(1000, lambda=5), nrow=200 )
col_cv(y)
n = dim(y)
colnames(y) = paste(rep("sample",n[[2]]),1:n[[2]])
rownames(y) = paste(rep("gene",n[[1]]),1:n[[1]])
y = as.data.frame(y)
col_cv(y)
}
