% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combat_fix.R
\name{combat_fix}
\alias{combat_fix}
\title{Title IT is exactly the same as sva::combat, only changing name. the sva website is
"https://bioconductor.org/packages/release/bioc/html/sva.html"}
\usage{
combat_fix(...)
}
\arguments{
\item{...}{the same as sva::combat}
}
\value{
matrix "?combat"
}
\description{
Title IT is exactly the same as sva::combat, only changing name. the sva website is
"https://bioconductor.org/packages/release/bioc/html/sva.html"
}
\examples{
#browseVignettes("sva")
}
