% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixbatch-package.R
\docType{package}
\name{fixbatch-package}
\alias{fixbatch}
\alias{fixbatch-package}
\title{fixbatch: The goal of packages is fixing the batch effect in data}
\description{
need add when i have time.
}
\author{
\strong{Maintainer}: JiaXuan Wang \email{poormouse@126.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
