% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeep.R
\name{housekeep}
\alias{housekeep}
\title{Title According to the coefficient of variation and the expression sum, the stable observation value is selected}
\usage{
housekeep(data, expcv = 1, exphigh = 1, explow = 0)
}
\arguments{
\item{data}{must be a data.frame or matrix}

\item{expcv}{a number range 0 to 1, Sets the maximum quantile value of the coefficient of variation}

\item{exphigh}{a number range 0 to 1, Sets the maximum quantile value of expression sum}

\item{explow}{a number range 0 to 1, Sets the minimum quantile value of expression sum}
}
\value{
data
}
\description{
Title According to the coefficient of variation and the expression sum, the stable observation value is selected
}
\examples{
y <- matrix( rpois(1000, lambda=5), nrow=200 )
n = dim(y)
colnames(y) = paste(rep("sample",n[[2]]),1:n[[2]])
rownames(y) = paste(rep("gene",n[[1]]),1:n[[1]])
y = as.data.frame(y)
#data_hk = housekeep(data, expcv = 0.3, exphigh = 0.8, explow = 0.4)
#head(data_hk)
}
