% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvca.R
\name{pvca}
\alias{pvca}
\title{A function for principal variance component analysis}
\usage{
pvca(counts, meta, threshold, inter)
}
\arguments{
\item{counts}{The Normalized(e.g. TMM)/ log-transformed reads count matrix from sequencing data (row:gene/feature, col:sample)}

\item{meta}{The Meta data matrix containing predictor variables (row:sample, col:predictor)}

\item{threshold}{The proportion of the variation in read counts explained by top k PCs. This value determines the number of top PCs to be used in pvca.}

\item{inter}{TRUE/FALSE - include/do not include pairwise interactions of predictors}
}
\value{
std.prop.val The vector of proportions of variation explained by each predictor.
}
\description{
The function is written based on the 'pvcaBatchAssess' function of the PVCA R package(http://watson.nci.nih.gov/bioc_mirror/packages/release/bioc/manuals/pvca/man/pvca.pdf).
and adjust by Donghyung Lee for changed slightly to make it more efficient and flexible for sequencing read counts data(https://github.com/dleelab/pvca). But it didn't update since 4 years ago.
My new R packages fixbatch need the function to estimate factor's partition of the total variability. So I fork the packages and Adjust in my own style.
}
