% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snm_fix.R
\name{snm_fix}
\alias{snm_fix}
\title{Title Extract the  normalizeed data by function "snm". detailed description in which
the function base on snm R packaegs(http://www.bioconductor.org/packages/release/bioc/html/snm.html),
packaged into fixbatch. Citation within R, enter citation("snm")).}
\usage{
snm_fix(...)
}
\arguments{
\item{...}{see in "?snm_fix_all" on "fixbatch"}
}
\value{
a normalizeed data
}
\description{
Title Extract the  normalizeed data by function "snm". detailed description in which
the function base on snm R packaegs(http://www.bioconductor.org/packages/release/bioc/html/snm.html),
packaged into fixbatch. Citation within R, enter citation("snm")).
}
\examples{
#library(bladderbatch)
#library(fixbatch)
#data(bladderdata)
#pheno = pData(bladderEset)
#edata = exprs(bladderEset)
#re1 = snm_fix_all(data = edata, batch = pheno,bv = 4, av = 3, iv = 2)
#re = snm_fix(data = edata, batch = pheno,bv = 4, av = 3, iv = 2)
}
