% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snm_fix.R
\name{snm_fix_all}
\alias{snm_fix_all}
\title{Title The function base on snm R packaegs(http://www.bioconductor.org/packages/release/bioc/html/snm.html), packaged into fixbatch. Citation within R, enter citation("snm")).}
\usage{
snm_fix_all(
  data = NULL,
  batch = NULL,
  bv = NULL,
  av = NULL,
  iv = NULL,
  bio.var = NULL,
  adj.var = NULL,
  int.var = NULL,
  diagnose = FALSE,
  verbose = TRUE,
  rm.adj = TRUE,
  ...
)
}
\arguments{
\item{data}{a matrix, Equivalent to raw.dat (?snm) , genes (rows) by sample (columns) matrix of unnormalized by other function.}

\item{batch}{a data frame, sample(rows) by variables(columns), including the biology variables,adjust}

\item{bv}{a number vector which column number is biology variables.}

\item{av}{a number vector which column number is adjust variables.}

\item{iv}{a number vector which column number is inter variables.}

\item{bio.var}{If bv is null, can use bio.var. A model matrix of the biological variable of interest. Required.}

\item{adj.var}{If av is null, can use adj.var. A model matrix of the adjustment variables.}

\item{int.var}{If iv is null, can use int.var. A data frame of the intensity-dependent adjustment variables, one effect per column.}

\item{diagnose}{If set to TRUE, then diagnostic plots are displayed for each iteration.}

\item{verbose}{If set to TRUE, then the iteration number is printed to the R console, and the above
plots are displayed for each iteration.}

\item{rm.adj}{This option is used to control whether or not the probe-specific adjustment variables
should be removed from the normalized data returned by the snm function call. If the analyst
will perform subsequent formal statistical inference (e.g., hypotheses testing), then this option
should be set to FALSE. If the analyst will perform clustering or network analysis, then this
option should be set to TRUE.}

\item{...}{see the parameter form "?snm" on snm R packages
i.e., the number of columns of raw.dat If bio.var contains an intercept, it is removed. If
adj.var=NULL, then an intercept is added automatically. It is possible not include int.var as
input, in which case it is treated as int.var=NULL and no intensity dependent effects are fit (see
examples below).}
}
\value{
a normalizeed s4 object
}
\description{
Title The function base on snm R packaegs(http://www.bioconductor.org/packages/release/bioc/html/snm.html), packaged into fixbatch. Citation within R, enter citation("snm")).
}
\examples{
#library(bladderbatch)
#library(fixbatch)
#data(bladderdata)
#pheno = pData(bladderEset)
#edata = exprs(bladderEset)
#re = snm_fix_all(data = edata, batch = pheno,bv = 4, av = 3, iv = 2)
}
